refactor appbar and sidebar into a wrapper component
